backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
markdown==3.7.
django-filter==25.1
django-cors-headers==4.7.0
mysqlclient==2.2.7
python-dotenv==1.1.0
djangorestframework-simplejwt==5.5.0
coverage==7.8.0